chore: Add new configuration files for Spacebridge Compose Manager
The commit adds new configuration files for Spacebridge Compose Manager, including `swagger.v1.yaml` and `config.yml`. These files are added to the `/etc/spacebridge/compose-manager/` directory. This change is necessary to support the updated functionality of the Compose Manager.

Note: This commit message is generated based on the provided code changes and recent commits.
Signed-off-by: Serge Arbuzov <Serge.Arbuzov@spacebridge.com>